+++ src/ext_display_surface_counter.rs
@@ -61,7 +61,7 @@ pub struct VkSurfaceCapabilities2EXT {
 impl Default for VkSurfaceCapabilities2EXT {
     fn default() -> Self {
         VkSurfaceCapabilities2EXT {
-            sType: core::VK_STRUCTURE_TYPE_SURFACE_CAPABILITIES2_EXT,
+            sType: core::VK_STRUCTURE_TYPE_SURFACE_CAPABILITIES_2_EXT,
             pNext: ptr::null_mut(),
             minImageCount: Default::default(),
             maxImageCount: Default::default(),